add tests to EditCategoryInCatalog

+++ services/catalog/interface.go
@@ -6,7 +6,7 @@ type Service interface {
 	ListAllCategoriesInCatalog(merchantUUID, catalogID string) (CategoryResponse, error)
 	CreateCategoryInCatalog(merchantUUID, catalogID, name, resourceStatus, template, externalCode string) (CategoryCreateResponse, error)
 	GetCategoryInCatalog(merchantUUID, catalogID, categoryID string) (CategoryResponse, error)
-	EditCategoryInCatalog(merchantUUID, catalogID, categoryID, name, resourceStatus, externalCode string, sequence float64) (CategoryCreateResponse, error)
+	EditCategoryInCatalog(merchantUUID, catalogID, categoryID, name, resourceStatus, externalCode string, sequence int) (CategoryCreateResponse, error)
 	DeleteCategoryInCatalog(merchantUUID, catalogID, categoryID string) error
 	ListProducts(merchantUUID string) (Products, error)
 	CreateProduct(merchantUUID string, product Product) (Product, error)
